Install Isolate binary and presere iso-codes pkgs
The worker image was intended to be built as a separate one, since
it relies on cgroups and kernel namespace to handle sandbox envs.
so it only can be run in "priviledge" mode only. Unfortunately,
it seems too early to split the worker from the core CMS image,
hence the main differences are the libcap, cgroups-mount and
isolate dependencies.

In the other hand, I found out that CMS services use iso-codes
files and mime type database at runtime so the Docker image must
preserve these non-compiled files.