Enhance env. variable conf. override
Months ago a python script was added to the CMS Docker images in
order to provide a configuration override through environment
variables imiatating some Spring conventions for such task, unfor-
tunately it turned to not be as convenient as initialy expected.

Because of this, the configuration override was redesigned to
directly map each configuration to an env. variable name instead
of trying to apply an generic approach. In other words, each of
configurations fields listed in the cms.conf and ranking.conf
files were mapped one by one with an specific env. variable name.