Expose env. var. mapping tables as tsv file
Instead of relying on correctly populating a python dictionary in
the script, the maps for each configuration and the env. variable
name that could override it were written on a tsv (tab separated)
file. I have also included the automation script that handles the
the transformation of the tsv file into a python dictionary right
into the getconfig.py file.